automatically train dicts/recompress
fixes #43, fixes #44

This might be too heavy handed. On start up, we'll scan dss_fetch_cache
for hosts that have at least 100 URLs but no dict. If we find any, we'll
create a dict and queue those entries to be recompressed.

On a sample db with 35K URLs, this dropped storage from 930MB to 430MB.
I suspect if we tuned the dictionaries we could do even better, but this
was ~free to implement (and surprisingly cheap at runtime, too)