perf note
This shaved another 10-15% in the fully-cached path. Once we have
experience with real life workloads, we can revisit if it makes
sense.

A related, more aggressive idea: the worker could optimistically
claim several URLs, and, upon hitting the first non-cached URL,
release any unprocessed ones.

I think the name of the game is to reduce the # of write transactions
per crawl so that we get 100% CPU utilization. When we have non 100%
CPU utilization, we must be blocked waiting for the write lock.