add extract-json-ld, @ sigil, __delete
The @ sigil is useful for cardinality many extractions,
e.g. links on a page, or JSON+LD entries. It creates an index
for fast deletions on future crawls.

Write lock contention starts to be a big deal--it might be good
to do SELECTs to see if the database actually needs to change
before rewriting things. That would also ensure that we don't
churn rowids needlessly.