avoid write locks
When extract_from_response returns data, check to see if it's already
in the DB. If it is, don't emit INSERT/DELETE statements.

There's room for more elision here. If run 1 emits:

```
delete @key=a
add @key=a value=foo
```

And run 2 emits:

```
delete @key=a
add @key=a value=foo
add @key=a value=bar
```

We could turn that into

```
add @key=a value=bar
```

We don't yet.